refactor (1) 사용하지 않는 메소드 삭제
