+++ src/main/java/ladder/domain/reward/Rewards.java
@@ -11,28 +11,15 @@ import java.util.stream.Collectors;
 
 public class Rewards {
     private final static String SPLIT_REGEX = ",";
-    private final static int MIN_REWARD_SIZE = 1;
     
     private final List<Reward> rewards;
     
-    private Rewards(int rewardSize, String rewardString) {
-        if (rewardSize < MIN_REWARD_SIZE) {
-            throw new IllegalArgumentException(ErrorMessages.CANT_INPUT_ZERO.message());
-        }
-        String[] rewards = rewardString.split(SPLIT_REGEX);
-        if (rewards.length != rewardSize) {
-            throw new IllegalArgumentException(ErrorMessages.NOT_MATCH_COUNT.message());
-        }
-        this.rewards = new ArrayList<>();
-        Arrays.stream(rewards).forEach(reward -> this.rewards.add(Reward.from(reward)));
+    public static Rewards newInstance() {
+        return new Rewards();
     }
     
     private Rewards() {
         this.rewards = new ArrayList<>();
-    }
-    
-    public static Rewards newInstance() {
-        return new Rewards();
     }
     
     public List<String> getRewardNames() {
@@ -41,12 +28,8 @@ public class Rewards {
           .collect(Collectors.toList());
     }
     
-    public int getSize() {
-        return rewards.size();
-    }
-    
     public String getReward(int i) {
-        if (i >= getSize()) {
+        if (i >= rewards.size()) {
             throw new IllegalArgumentException(ErrorMessages.OVER_INPUT_REWARD.message());
         }
         return rewards.get(i).getReward();
