+++ src/main/java/ladder/domain/reward/info/Reward.java
@@ -6,15 +6,15 @@ import ladder.util.StringUtils;
 public class Reward {
     private final String reward;
     
+    public static Reward from(String reward) {
+        return new Reward(reward);
+    }
+    
     private Reward(String reward) {
         if (StringUtils.isEmpty(reward)) {
             throw new IllegalArgumentException(ErrorMessages.CANT_INPUT_EMPTY_STRING.message());
         }
         this.reward = reward;
-    }
-    
-    public static Reward from(String reward) {
-        return new Reward(reward);
     }
     
     public String getReward() {
