+++ src/main/java/ladder/view/component/constant/Step.java
@@ -1,4 +1,4 @@
-package ladder.view.constant;
+package ladder.view.component.constant;
 
 import java.util.Arrays;
 
@@ -11,27 +11,16 @@ public enum Step {
     GAMER_NAME_INPUT_STEP(5),
     REWARD_STEP(6);
     
-    public static Step thisStep = GAMERS_STEP;
+    public static Step THIS_STEP = GAMERS_STEP;
     
     private int step;
     
     public static boolean isThisStep(Step step) {
-        return thisStep == step;
+        return THIS_STEP == step;
     }
     
     public static void setNextStep(Step step) {
-        thisStep = step;
-    }
-    
-    private static Step getNextStep(Step step) {
-        return findStep(step.getStep() + 1);
-    }
-    
-    private static Step findStep(int i) {
-        return Arrays.stream(values())
-          .filter(step -> step.getStep() == i)
-          .findFirst()
-          .orElse(EMPTY_STEP);
+        THIS_STEP = step;
     }
     
     Step(int step) {
