+++ src/main/java/ladder/view/component/gamer/GamerView.java
@@ -7,7 +7,7 @@ import ladder.core.view.ViewImpl;
 import ladder.core.view.input.Inputor;
 import ladder.core.view.output.Printer;
 import ladder.view.component.View;
-import ladder.view.constant.Step;
+import ladder.view.component.constant.Step;
 
 public class GamerView implements ViewImpl {
     private final static String ANSWER = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
@@ -28,8 +28,8 @@ public class GamerView implements ViewImpl {
         if (!Step.isThisStep(view.getStep())) {
             return;
         }
-        view.print(ANSWER);
         Step.setNextStep(Step.REWARD_INPUT_STEP);
+        view.print(ANSWER);
         controller.inputGamers(view.inputString());
     }
 }
