+++ src/main/java/ladder/view/component/ladder/LadderView.java
@@ -6,7 +6,7 @@ import ladder.core.view.ViewImpl;
 import ladder.core.view.input.Inputor;
 import ladder.core.view.output.Printer;
 import ladder.view.component.View;
-import ladder.view.constant.Step;
+import ladder.view.component.constant.Step;
 
 public class LadderView implements ViewImpl {
     private static final String ANSWER = "최대 사다리 높이는 몇 개인가요?";
@@ -27,8 +27,8 @@ public class LadderView implements ViewImpl {
         if (!Step.isThisStep(view.getStep())) {
                  return;
         }
-        view.print(ANSWER);
         Step.setNextStep(Step.RESULT_STEP);
+        view.print(ANSWER);
         controller.inputLadderSize(view.inputNumber());
     }
 }
